refactor(common): move `prisma.services` into `common/services/prisma`
