+++ src/regency/regency.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PrismaService } from '~/src/prisma.service';
+import { PrismaService } from '~/src/common/services/prisma';
 import { RegencyController } from './regency.controller';
 import { RegencyService } from './regency.service';
