+++ src/village/village.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { Village } from '@prisma/client';
 import { SortHelper, SortOptions } from '~/src/common/helper/sort';
-import { PrismaService } from '~/src/prisma.service';
+import { PrismaService } from '~/src/common/services/prisma';
 
 type VillageSortKeys = keyof Village;
